Fix type annotation on keyword argument in `copy(**add_or_replace)`
The type annotation for keyword arguments describes the type of the  _value_, not of the kwargs as a whole itself. (See https://www.python.org/dev/peps/pep-0484/#id37)
This is because the keys of a keyword argument paramater is always `str`.

So the signature of of `copy(self, **add_or_replace: Dict[_K, _V])` means "add_or_replace is a keyword arguments parameter whose _values_ are of type `Dict[_K, _V]`", probably not what was intended.

Incidentally, the return type annotation is incorrect if the self `immutabledict` didn't have `Text` keys or a supertype thereof, but that is a separate issue.